refactor: 북마크 관련 파일들을 bookmark 패키지 아래로 이동
