+++ src/main/java/com/techfork/domain/activity/bookmark/dto/BookmarkDto.java
@@ -1,0 +1,23 @@
+package com.techfork.domain.activity.bookmark.dto;
+
+import lombok.Builder;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Builder
+public record BookmarkDto(
+        Long bookmarkId,
+        Long postId,
+        String title,
+        String shortSummary,
+        String url,
+        String companyName,
+        String logoUrl,
+        LocalDateTime publishedAt,
+        String thumbnailUrl,
+        Long viewCount,
+        List<String> keywords,
+        Boolean isBookmarked
+) {
+}
