+++ src/main/java/com/techfork/domain/activity/controller/ActivityController.java
@@ -2,8 +2,8 @@ package com.techfork.domain.activity.controller;
 
 import com.techfork.domain.activity.bookmark.service.BookmarkCommandService;
 import com.techfork.domain.activity.bookmark.service.BookmarkQueryService;
-import com.techfork.domain.activity.dto.BookmarkListResponse;
-import com.techfork.domain.activity.dto.BookmarkRequest;
+import com.techfork.domain.activity.bookmark.dto.BookmarkListResponse;
+import com.techfork.domain.activity.bookmark.dto.BookmarkRequest;
 import com.techfork.domain.activity.dto.ReadPostListResponse;
 import com.techfork.domain.activity.dto.ReadPostRequest;
 import com.techfork.domain.activity.dto.SearchHistoryRequest;
@@ -33,7 +33,7 @@ public class ActivityController {
     private final ActivityQueryService activityQueryService;
     private final BookmarkCommandService bookmarkCommandService;
     private final BookmarkQueryService bookmarkQueryService;
-    
+
     @Operation(
             summary = "읽은 게시글 목록 조회",
             description = "사용자가 읽은 게시글 목록을 조회합니다. 최근 읽은 순서로 정렬됩니다."
@@ -49,7 +49,7 @@ public class ActivityController {
         ReadPostListResponse response = activityQueryService.getReadPosts(userPrincipal.getId(), lastReadPostId, size);
         return BaseResponse.of(SuccessCode.OK, response);
     }
-    
+
     @Operation(
             summary = "읽은 게시글 저장",
             description = "사용자가 특정 게시글을 읽은 기록을 저장합니다. 읽은 시간과 체류 시간을 기록합니다."
