--- src/main/java/com/techfork/domain/activity/dto/BookmarkListResponse.java
@@ -1,13 +1,0 @@
-package com.techfork.domain.activity.dto;
-
-import lombok.Builder;
-
-import java.util.List;
-
-@Builder
-public record BookmarkListResponse(
-        List<BookmarkDto> bookmarks,
-        Long lastBookmarkId,
-        boolean hasNext
-) {
-}
