--- src/main/java/com/techfork/domain/activity/dto/BookmarkRequest.java
@@ -1,9 +1,0 @@
-package com.techfork.domain.activity.dto;
-
-import jakarta.validation.constraints.NotNull;
-
-public record BookmarkRequest(
-        @NotNull(message = "게시글 ID는 필수입니다.")
-        Long postId
-) {
-}
