+++ src/test/java/com/techfork/domain/post/controller/PostControllerIntegrationTest.java
@@ -1,7 +1,7 @@
 package com.techfork.domain.post.controller;
 
-import com.techfork.domain.activity.entity.Bookmark;
-import com.techfork.domain.activity.repository.BookmarkRepository;
+import com.techfork.domain.activity.bookmark.entity.Bookmark;
+import com.techfork.domain.activity.bookmark.repository.BookmarkRepository;
 import com.techfork.domain.post.entity.Post;
 import com.techfork.domain.post.entity.PostKeyword;
 import com.techfork.domain.post.repository.PostKeywordRepository;
